\name{getChrs}
\alias{getChrs}

\title{
Chromosome ID
}

\description{
List the chromosome ID
}

\usage{
getChrs()
}

\details{
List the chromosome ID supported by the package analysis
}

\value{
A character vector of chromosome ID
}

\references{
}

\author{
H Mei; L Li
}

\note{
}

\seealso{
}

\examples{
getChrs()
}

\keyword{ chromosome }

