\name{T2DGWES}
\alias{T2DGWES}
\docType{data}
\title{
Sample T2D-GWES Data
}
\description{
Genome-wide expression study (GWES) of differential gene expression at pancreases of 10 control 
and 10 T2D human subjects    
}

\usage{
data(T2DGWES)
}

\format{
	
}

\details{ 
	The 'T2DGWES' data loads a data frame of variable, "T2DExpression", which shows the
	differential gene expression p-values for T2D and control subjects. The data frame contains
	20,185 genes with gene symbol, Entrez gene ID and differential p-value. The p-values 
 	were obtained by the method of linear models with empirical Bayes adjusting.   	
	
}

\source{
}

\references{
Marselli L, Thorne J, Dahiya S, Sgroi DC, Sharma A, Bonner-Weir S, Marchetti P, Weir GC: Gene 
expression profiles of Beta-cell enriched tissue obtained by laser capture microdissection 
from subjects with type 2 diabetes. PLoS One 2010, 5(7):e11499
}

\examples{
}
\keyword{datasets}
\keyword{permutation}
